Fix broken imports and undefined variables in confection refactor
- Fix fill_config returning undefined 'defaulted' instead of 'overrided'
- Add missing imports: warnings, Dict, List in _interpolation.py
- Replace srsly.json_loads/json_dumps with stdlib json in util.py
- Add missing ConfigValidationError import in util.py
- Create _constants.py module for shared constants (VARIABLE_RE,
  SECTION_PREFIX, ARGS_FIELD, etc.) to avoid circular imports
- Fix __init__.py importing symbols from wrong modules
