+++ confection/__init__.py
@@ -1,14 +1,13 @@
 # FIXME some symbols are not in __all__; can we remove them?
-from ._config import (  # noqa: F401
+from ._config import Config  # noqa: F401
+from ._constants import (  # noqa: F401
     ARGS_FIELD,
     ARGS_FIELD_ALIAS,
     RESERVED_FIELDS,
     SECTION_PREFIX,
     VARIABLE_RE,
-    Config,
-    try_dump_json,
-    try_load_json,
 )
+from .util import try_dump_json, try_load_json  # noqa: F401
 from ._errors import ConfigValidationError
 from ._registry import Promise, registry
 from .util import SimpleFrozenDict, SimpleFrozenList
