+++ confection/util.py
@@ -1,12 +1,13 @@
 import functools
+import json
+import re
 from copy import deepcopy
-from typing import Any, Callable, Iterator, Protocol, TypeVar
+from typing import Any, Callable, Protocol, TypeVar
 
-from pydantic import GetCoreSchemaHandler
-from pydantic_core import core_schema
+from ._constants import VARIABLE_RE
+from ._errors import ConfigValidationError
 
 _DIn = TypeVar("_DIn")
-
 
 class Decorator(Protocol):
     """Protocol to mark a function as returning its child with identical signature."""
@@ -27,28 +28,6 @@ def partial(
     partial_func = functools.partial(func, *args, **kwargs)
     partial_func.__doc__ = func.__doc__
     return partial_func
-
-
-class Generator(Iterator):
-    """Custom generator type. Used to annotate function arguments that accept
-    generators so they can be validated by pydantic (which doesn't support
-    iterators/iterables otherwise).
-    """
-
-    @classmethod
-    def __get_pydantic_core_schema__(
-        cls,
-        _source_type: Any,
-        _handler: GetCoreSchemaHandler,
-    ) -> core_schema.CoreSchema:
-        return core_schema.with_info_plain_validator_function(cls.__validate__)
-
-    @classmethod
-    def __validate__(cls, v, info):
-        if not hasattr(v, "__iter__") and not hasattr(v, "__next__"):
-            raise TypeError("not a valid iterator")
-        else:
-            return v
 
 
 DEFAULT_FROZEN_DICT_ERROR = (
@@ -150,3 +129,34 @@ def is_promise(obj) -> bool:
     if len(id_keys):
         return True
     return False
+
+
+def try_load_json(value: str) -> Any:
+    """Load a JSON string if possible, otherwise default to original value."""
+    try:
+        return json.loads(value)
+    except Exception:
+        return value
+
+
+def try_dump_json(value: Any, data: dict[str, dict] | str = "") -> str:
+    """Dump a config value as JSON and output user-friendly error if it fails."""
+    # Special case if we have a variable: it's already a string so don't dump
+    # to preserve ${x:y} vs. "${x:y}"
+    if isinstance(value, str) and VARIABLE_RE.search(value):
+        return value
+    try:
+        value = json.dumps(value)
+    except Exception as e:
+        err_msg = (
+            f"Couldn't serialize config value of type {type(value)}: {e}. Make "
+            f"sure all values in your config are JSON-serializable. If you want "
+            f"to include Python objects, use a registered function that returns "
+            f"the object instead."
+        )
+        raise ConfigValidationError(config=data, desc=err_msg) from e
+    # Escape $ to $$ for configparser, but preserve ${...} variable references
+    return re.sub(r"\$(?!\{)", "$$", value)
+
+
+
