+++ tests/test_registry.py
